fix dcgan

+++ vision/transforms/normalize.go
@@ -9,11 +9,11 @@ import (
 // NormalizeTransformer corresponds to torchvision.transforms.html#Normalize. It
 // implements Go interface gotorch/data.Transform.
 type NormalizeTransformer struct {
-	Mean, Stddev []float64
+	Mean, Stddev []float32
 }
 
 // Normalize returns normalize transformer
-func Normalize(mean []float64, stddev []float64) *NormalizeTransformer {
+func Normalize(mean []float32, stddev []float32) *NormalizeTransformer {
 	return &NormalizeTransformer{mean, stddev}
 }
 
@@ -22,16 +22,16 @@ func (t *NormalizeTransformer) Run(input torch.Tensor) torch.Tensor {
 	var meanT torch.Tensor
 	var stddevT torch.Tensor
 	if len(t.Mean) == 1 {
-		meanT = torch.NewTensor([][][]float64{{{t.Mean[0]}}})
+		meanT = torch.NewTensor([][][]float32{{{t.Mean[0]}}})
 	} else if len(t.Mean) == 3 {
-		meanT = torch.NewTensor([][][]float64{{{t.Mean[0]}}, {{t.Mean[1]}}, {{t.Mean[2]}}})
+		meanT = torch.NewTensor([][][]float32{{{t.Mean[0]}}, {{t.Mean[1]}}, {{t.Mean[2]}}})
 	} else {
 		panic(fmt.Sprintf("len(Mean) should be 1 or 3."))
 	}
 	if len(t.Stddev) == 1 {
-		stddevT = torch.NewTensor([][][]float64{{{t.Stddev[0]}}})
+		stddevT = torch.NewTensor([][][]float32{{{t.Stddev[0]}}})
 	} else if len(t.Stddev) == 3 {
-		stddevT = torch.NewTensor([][][]float64{{{t.Stddev[0]}}, {{t.Stddev[1]}}, {{t.Stddev[2]}}})
+		stddevT = torch.NewTensor([][][]float32{{{t.Stddev[0]}}, {{t.Stddev[1]}}, {{t.Stddev[2]}}})
 	} else {
 		panic(fmt.Sprintf("len(Stddev) should be 1 or 3."))
 	}
